requirements.txt
flask
flask-cors
python-pptx
gunicorn